修改 直属下级列表 接口

+++ job-executor/src/main/java/com/job/executor/domain/AgentCommission.java
@@ -42,7 +42,7 @@ public class AgentCommission {
     /**
      * 直属业绩
      */
-    private Long subRatio;
+    private BigDecimal subRatio;
 
     /**
      * 直属佣金
@@ -57,7 +57,7 @@ public class AgentCommission {
     /**
      * 下属业绩
      */
-    private Long otherRatio;
+    private BigDecimal otherRatio;
 
     /**
      * 下属佣金
